Translation: Log a warning for any unknown lang keys
Based on https://laravel.com/docs/10.x/localization#handling-missing-translation-strings.

Use "resolving" callback instead of using the Log facade / the 'log'
service because if we do that it will force the deferred
LogServiceProvider to be registered on every request, essentially
making it not deferred any more.

Similar to how Model::shouldBeStrict() is only called when not
in production, we only throw an exception when not in production.