ESLint: Change import/extensions from "always" to "ignorePackages"
It turns out set to "always" is a pain, for example if we write:

    import { createRoot } from "react-dom/client"

import/extensions doesn't like this. So if you add the explicit
file extension:

    import { createRoot } from "react-dom/client.js"

That makes ESLint happy, but tsc now fails because it can't find
the types definition for it, because @types/react-dom only
has "exports" defined for "/client" not "/client.js".

So it seems its impossible to enforce extensions on package imports
while at the same time using tsc.